requirements.txt
tensorflow
statsmodels
scikit-learn
keras-tuner
seaborn
matplotlib
pandas[performance]